+++ ATVO.ThemesSDK/Data/Lap/ISector.cs
@@ -4,7 +4,7 @@ namespace ATVO.ThemesSDK.Data.Lap
 {
     public interface ISector : INotifyPropertyChanged
     {
-        byte Index { get; }
+        int Index { get; }
         float Time { get; }
         int ReplayPosition { get; }
     }
